oh-tab-b7qq: extract dev bridge logger
